Remove the is_active=1 filter for wikibase.cloud import
I ran the metadata import, and it imported a bunch of wikibase.cloud
stuff.
I can only assume this is because it contains the ones not flagged as active.
As we are going to import them from there, we might as well import
them from the "real" place in the end, ie directly from wikibase.cloud